requirements.txt
openai>=1.30.0
anthropic>=0.25.0
pandas>=2.0.0
python-dotenv>=1.0.0
tqdm>=4.66.0
